--- config/tests/test_middleware.py
@@ -1,6 +1,0 @@
-from django.conf import settings
-
-
-def test_SSODisplayLoggedInCookieMiddleware_installed():
-    expected = 'config.middleware.SSODisplayLoggedInCookieMiddleware'
-    assert expected in settings.MIDDLEWARE_CLASSES
