+++ core/tests/test_middleware.py
@@ -1,0 +1,3 @@
+def test_SSODisplayLoggedInCookieMiddleware_installed(settings):
+    expected = 'core.middleware.SSODisplayLoggedInCookieMiddleware'
+    assert expected in settings.MIDDLEWARE_CLASSES
