+++ docker/cmd-integration-test-webserver.sh
@@ -5,4 +5,4 @@ python /usr/src/app/manage.py create_test_users
 python /usr/src/app/manage.py loaddata test_fixtures/load_tests.json
 mkdir -p /usr/src/app/sso/static
 python /usr/src/app/manage.py collectstatic --noinput
-gunicorn config.wsgi --bind 0.0.0.0:$PORT --log-file -
+gunicorn cong.wsgi --bind 0.0.0.0:$PORT --log-file -
