+++ manage.py
@@ -4,7 +4,7 @@ import sys
 
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "conf.settings")
 
     from django.core.management import execute_from_command_line
