+++ requirements.txt
@@ -12,7 +12,7 @@ contextlib2==0.5.5        # via raven
 defusedxml==0.5.0         # via python3-openid
 directory-api-external==4.2.1
 directory-client-core==0.2.1  # via directory-api-external
-directory-components==2.6.0
+directory-components==2.7.1
 directory-constants==6.5.2
 directory-healthcheck==0.4.2
 directory-validators==4.5.1
