+++ sso/healthcheck/views.py
@@ -1,7 +1,7 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from config.signature import SignatureCheckPermission
+from conf.signature import SignatureCheckPermission
 
 from directory_healthcheck.views import BaseHealthCheckAPIView
